Fixed pow ambiguity.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@1212 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/serial/objistrasn.cpp
@@ -30,6 +30,9 @@
 *
 * ---------------------------------------------------------------------------
 * $Log$
+* Revision 1.33  2000/01/11 14:16:45  vasilche
+* Fixed pow ambiguity.
+*
 * Revision 1.32  2000/01/10 19:46:40  vasilche
 * Fixed encoding/decoding of REAL type.
 * Fixed encoding/decoding of StringStore.
@@ -601,7 +604,7 @@ double CObjectIStreamAsn::ReadDouble(void)
     		return 0;
     }
 
-	return im * pow(base, ic);
+	return im * pow(double(base), ic);
 }
 
 void CObjectIStreamAsn::ReadString(string& s)
